Refactor FileNavigator and DirectoryView to take advantage of new ListSelect widget. Closes #759.
